Fix static analysis issues identified during Clang 20 upgrade
- Fix signed/unsigned comparison in ZigGlobalObject.cpp stack trace iteration
- Add bounds checking to FFI callback functions to prevent buffer overflows
- Document clang-analyzer crash with JavaScriptCore headers compatibility issue

These fixes address potential security vulnerabilities and improve code robustness.

+++ src/bun.js/bindings/ZigGlobalObject.cpp
@@ -700,7 +700,7 @@ static JSValue computeErrorInfoWithPrepareStackTrace(JSC::VM& vm, Zig::GlobalObj
 
     // We need to sourcemap it if it's a GlobalObject.
 
-    for (int i = 0; i < stackTrace.size(); i++) {
+    for (size_t i = 0; i < stackTrace.size(); i++) {
         ZigStackFrame frame = {};
         auto& stackFrame = stackFrames.at(i);
         String sourceURLForFrame = Zig::sourceURL(vm, stackFrame);
